feat: add ignore_proxied config support
Adding this change will stop the `proxied` setting to be forcefully set for any `_PROXIABLE_RECORD_TYPES` record.

Currently, if the `source` DNS doesn't support `proxied` setting, instead of ignoring the **proxied** flag, it will always be set to `false` overriding a change in the Cloudflare's DNS interface. 

IMHO, if the `proxied` flag is not set for a given record, it should not be set to `false`  but ignored.