Ensure unsupported record types are handled correctly
During the apply update phase, all zone records are iterated and updated
as required, regardless of whether the record type is supported by the
provider or not. An `AttributeError` exception is triggered if any
unsupported type exists within the zone, regardless of the actual record
type being updated. This problem appears to only trigger for zones with
existing records that have been onboarded onto OctoDNS (eg: DS records
existing within the zone). Attempts to use the
`octodns.processor.filter.TypeRejectlistFilter` processor to ignore the
offending type was of no help due to the way the methods are named
(`_data_for_{_type}`) within this provider.

       Traceback (most recent call last):
         ...
         File "./venv/lib/python3.12/site-packages/octodns_cloudflare/__init__.py", line 911, in _apply_Update
           r = self._record_for(zone, name, record['type'], [record], True)
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
         File "./venv/lib/python3.12/site-packages/octodns_cloudflare/__init__.py", line 435, in _record_for
           data_for = getattr(self, f'_data_for_{_type}')
                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
       AttributeError: 'CloudflareProvider' object has no attribute '_data_for_DS'. Did you mean: '_data_for_NS'?
       exit status 1

The check for supported types has been relocated into `zone_records` to
centralize the `SUPPORTS` logic. Unsupported DNS records are skipped.