Add option to retry on auth errors
Sometimes the Cloudflare API will return a 403 response for a DNS record
change request for seemingly no good reason. This behavior is difficult
to reproduce reliably, as it seems to only happen when a large number
(hundreds) of zones needs to be created from scratch.

This change makes it possible to create a large amount of zones without
having to manually restart the process after random 403 responses.