Add option to log name server information on apply
Enable logging zone name server information on apply. This makes
managing new zones easier, as otherwise one has to go look up the name
server information for the zone separately in order to configure the domain
registrar side records.

The name server information is already present in the responses that the
zone method receives, so adding it is easy enough. Downside is that a
nested dict structure is required, which makes lookups a bit more
complex.